Fix a large amount of warnings:  * Integer signed with unsigned comparisons  * "/*" found within comments: instead use "#if 0" sections to comment out code (as that approach nests properly)  * Remove unused variables  * Use iterators to iterate over containers, not indexes  * Mark a location where performance optimisations might be possible
git-svn-id: https://svn.il.fontys.nl/usersvn/fishtank/svn/trunk@72 bb1b31de-baa9-489e-af1c-03370f2903b8

+++ Vissen Project Aquarium/imagereceiver.h
@@ -61,8 +61,8 @@ class ImageReceiver
 	int m_server_socket;
 	SimpleSocketStream m_socket_stream;
 	std::vector<unsigned char> buffer;/// contains image
-	int buffered_bytes;
-	int image_size;
+	size_t buffered_bytes;
+	size_t image_size;
 
 	void AcceptClient();
 	bool ReceiveSegment();
